Update README.md
In Java and array and array list have a pretty similar function; to store a collection of elements.


Although they have a similar function their characteristic is different.

An Array is a fixed-size collection of elements of the same data type (fixed length). Once the elements are set they cannot be changed.

The array cannot change size at run time. The programmer may have to guess what the number of elements may be required for the array.

Examples of declaring an array:

double[] values = new double[10]; the values variable holds a reference to the array (the reference denotes the location of the array in memory)

double[] moreValues = { 1, 2, 3, 4, 5 };



The array-list is part of the Java Collections Framework and is a dynamic array-like data structure that can store and manipulate a collection of elements. (Not fixed length).

The array-list will grow or shrink in size dynamically as elements are added or removed.

To use the arraylist, you need to import java.util.Arraylist.

Example of declaring and arraylist:

ArrayList<String> friends = new ArrayList<String>();

ArrayList<String> -----> Variable type

friends ------> variable name 

new ArrayList<String>() --------> An arrya list of object size 0